Extract `after_test` helper
